+++ small_text/classifiers/factories.py
@@ -15,6 +15,14 @@ class AbstractClassifierFactory(ABC):
 class SklearnClassifierFactory(AbstractClassifierFactory):
 
     def __init__(self, base_estimator, num_classes, kwargs={}):
+        """
+        base_estimator : BaseEstimator
+            A scikit learn estimator which is used as a template for creating new classifier objects.
+        num_classes : int
+            Number of classes.
+        kwargs : dict
+            Keyword arguments that are passed to the constructor of each classifier that is built by the factory.
+        """
         if not issubclass(type(base_estimator), BaseEstimator):
             raise ValueError(
                 'Given classifier template must be a subclass of '
@@ -27,6 +35,13 @@ class SklearnClassifierFactory(AbstractClassifierFactory):
         self.kwargs = kwargs
 
     def new(self):
+        """Creates a new SklearnClassifier instance.
+
+        Returns
+        -------
+        classifier : SklearnClassifier
+            A new instance of SklearnClassifier which is initialized with the given keyword args `kwargs`.
+        """
         return SklearnClassifier(clone(self.base_estimator), self.num_classes, **self.kwargs)
 
     def __str__(self):